+++ src/main/kotlin/org/ua/wohnung/bot/flows/dto/ChatMetadata.kt
@@ -1,8 +1,8 @@
 package org.ua.wohnung.bot.flows.dto
 
 data class ChatMetadata(
-    val userId: Long,
-    val chatId: Long,
+    val userId: Int,
+    val chatId: Int,
     val username: String?,
     val input: String
 )
